We can navigate to the image even if multiple images are uploaded to the application with the same title.
Earlier the users were unable to navigate to the image if multiple images are uploaded to system with same title. The used to throw runtime error. Now we changed the business logic so that the user can easily navigate to any particular image even if there are mulitple images with same title